backend/requirements.txt
google-generativeai
pymupdf
python-dotenv
langextract
pandas
reportlab
fastapi
uvicorn
python-multipart
aiofiles
